Move App component into components direcotry, edit imports
