+++ src/components/App/App.scss
@@ -1,4 +1,4 @@
-@import './partials/variables';
+@import '../../partials/variables';
 
 html {
   height: 100%;
