+++ src/index.scss
@@ -1,10 +1,10 @@
 
 @import "../node_modules/normalize-scss/sass/normalize";
 @import url('https://fonts.googleapis.com/css2?family=Major+Mono+Display&family=Poppins:ital,wght@0,200;0,400;0,500;1,200&display=swap');
-@import './partials/variables',
-'./partials/extends',
-'./partials/mixins',
-'./App.scss',
+@import './partials/variables.scss',
+'./partials/extends.scss',
+'./partials/mixins.scss',
+'./components/App/App.scss',
 './components/SongCard/SongCard.scss',
 './components/Navigation/Navigation.scss',
 './components/SearchBar/SearchBar.scss',
